disable iframe component helper
#131

too many gotchas.  This should really be done as a markdown extension (hugo shortcode?) but there aren't any great plugins for this... doing it after markdown causes all sorts of issues (e.g. markdown will flag the internal url and convert it to a hyperlink; or typography changes the "" to something prettier, but it breaks the rexexp...